refactor(dh): rename function
