+++ libs/dh/metering-point/domain/src/index.ts
@@ -14,4 +14,4 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export * from './lib/is-metering-point-id';
+export * from './lib/is-valid-metering-point-id';
