+++ libs/dh/metering-point/domain/src/lib/is-valid-metering-point-id.ts
